refactor(soundxyz): move abi to constants file
